agentic-financial-advisor/requirements.txt
pyautogen
yfinance
pandas
numpy
matplotlib
